nix/requirements.txt
requests==2.11.1
bs4==0.0.1
Django==1.10.1
djangorestframework==3.4.7
celery==3.1.25
redis==2.10.5
django-celery==3.2.1

